requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.10
pydantic==2.7.1
pydantic-core==2.18.2
python-dotenv==1.0.1
gunicorn==22.0.0
sqladmin==0.16.1